Break out utils methods
