+++ package.json
@@ -47,7 +47,7 @@
     "jest": "^20.0.4",
     "jest-cli": "^20.0.4",
     "nodemon": "^1.11.0",
-    "rewire": "^2.5.2"
+    "supertest": "^3.3.0"
   },
   "jest": {
     "coverageDirectory": ".coverage",
